requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.0.4
dataclasses==0.6
Flask==2.0.3
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-http-client==3.3.7
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==3.141.0
sendgrid==6.9.7
six==1.16.0
soupsieve==2.3.2.post1
starkbank-ecdsa==2.2.0
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.0.3
zipp==3.6.0
webdriver-manager
